backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
mysql-connector-python>=8.2.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
